requirements.txt
Django==3.2.12
djangorestframework==3.13.1
psycopg2-binary==2.9.3
python-decouple==3.6
dj-database-url==0.5.0
pyyaml==6.0
uritemplate==4.1.1
gunicorn==20.1.0
whitenoise==6.0.0
